+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `CreateDatabaseTimeoutInSeconds` static property to `DiscoveredServerHelper`
 
 ## [1.0.7] - 2021-05-18